+++ migrations/m250422_145411_default_template.php
@@ -1,0 +1,22 @@
+<?php
+
+use humhub\components\Migration;
+
+class m250422_145411_default_template extends Migration
+{
+    /**
+     * {@inheritdoc}
+     */
+    public function safeUp()
+    {
+        $this->safeAddColumn('custom_pages_template', 'is_default', $this->boolean()->notNull()->defaultValue(0));
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function safeDown()
+    {
+        $this->safeDropColumn('custom_pages_template', 'is_default');
+    }
+}