+++ modules/template/controllers/ContainerContentController.php
@@ -193,7 +193,7 @@ class ContainerContentController extends Controller
         $form->setItemTemplate($itemTemplate);
         $form->setScenario('edit');
 
-        if (Yii::$app->request->post() && $form->load(Yii::$app->request->post()) && $form->save()) {
+        if ($form->load(Yii::$app->request->post()) && $form->save()) {
             TemplateCache::flushByElementContent($elementContent);
             return $this->asJson([
                 'success' => true,