+++ modules/template/services/ImportService.php
@@ -8,24 +8,35 @@
 
 namespace humhub\modules\custom_pages\modules\template\services;
 
+use humhub\modules\custom_pages\Module;
+use humhub\modules\custom_pages\modules\template\events\DefaultTemplateEvent;
 use humhub\modules\custom_pages\modules\template\models\Template;
 use humhub\modules\custom_pages\modules\template\models\TemplateElement;
 use humhub\modules\file\models\FileContent;
-use yii\base\InvalidConfigException;
 use Yii;
 use yii\db\ActiveRecord;
 
 class ImportService
 {
-    private string $type;
-    private string $filePath;
+    public const EVENT_DEFAULT_TEMPLATES = 'defaultTemplates';
+
+    private ?string $type = null;
     private array $errors = [];
     public ?Template $template = null;
+    public bool $allowUpdateDefaultTemplates = false;
 
-    public function __construct(string $type, string $filePath)
+    public function __construct(?string $type = null)
     {
         $this->type = $type;
-        $this->filePath = $filePath;
+
+        if ($module = $this->getModule()) {
+            $this->allowUpdateDefaultTemplates = $module->allowUpdateDefaultTemplates;
+        }
+    }
+
+    public static function instance(?string $type = null): self
+    {
+        return new self($type);
     }
 
     public function addError(string $error): void
@@ -43,15 +54,32 @@ class ImportService
         return $this->errors;
     }
 
-    public function run(): bool
+    public function importFromFolder(string $path): bool
     {
-        if (!file_exists($this->filePath)) {
+        if (!is_dir($path)) {
+            $this->addError('Wrong default templates path "' . $path . '"!');
+            return false;
+        }
+
+        $result = true;
+        foreach (scandir($path) as $file) {
+            if (str_ends_with($file, '.json')) {
+                $result = $this->importFromFile($path . '/' . $file) && $result;
+            }
+        }
+
+        return $result;
+    }
+
+    public function importFromFile(string $path): bool
+    {
+        if (!file_exists($path)) {
             $this->addError('The import file is not found!');
             return false;
         }
 
         try {
-            $data = json_decode(file_get_contents($this->filePath), true);
+            $data = json_decode(file_get_contents($path), true);
         } catch (\Exception $e) {
             $this->addError('The import file is not readable! Error: ' . $e->getMessage());
             return false;
@@ -62,21 +90,37 @@ class ImportService
             return false;
         }
 
-        if (isset($data['type']) && $data['type'] !== $this->type) {
+        if (isset($data['type'], $this->type) && $data['type'] !== $this->type) {
             $this->addError(Yii::t('CustomPagesModule.template', 'The template can be imported only as {type}!', [
                 'type' => Template::getTypeTitle($data['type']),
             ]));
             return false;
         }
 
+        return $this->runImport($data);
+    }
+
+    public function runImport(array $data): bool
+    {
         if (!$this->importTemplate($data)) {
             return false;
         }
 
+        $importedElementNames = [];
         if (isset($data['elements']) && is_array($data['elements'])) {
             foreach ($data['elements'] as $element) {
-                $this->importElement($element);
+                if ($element = $this->importElement($element)) {
+                    $importedElementNames[] = $element->name;
+                }
             }
+        }
+
+        // Delete old template elements that are not found in the new template
+        $oldElements = TemplateElement::find()
+            ->where(['template_id' => $this->template->id])
+            ->andWhere(['NOT IN', 'name', $importedElementNames]);
+        foreach ($oldElements->each() as $oldElement) {
+            $oldElement->delete();
         }
 
         return !$this->hasErrors();
@@ -94,22 +138,38 @@ class ImportService
 
     private function importTemplate(array $data): bool
     {
-        $template = Template::findOne(['name' => $data['name']]);
-        if ($template instanceof Template) {
-            // Delete old template and its linked elements before importing a new one with the same name
-            if (!$template->delete()) {
-                $this->addError('Cannot delete the old template "' . $data['name'] . '"!');
+        $template = Template::findOne(['name' => $data['name']]) ?? new Template();
+
+        if ($template->is_default && !$template->isNewRecord) {
+            // Check if default templates can be updated
+            if (!$this->allowUpdateDefaultTemplates) {
+                $this->addError(Yii::t('CustomPagesModule.template', 'Cannot import default template!'));
                 return false;
+            }
+
+            // If the default template was modified
+            if ($template->updated_at !== null || $template->updated_by !== null) {
+                // Rename the modified default template
+                $uniqueName = $template->name . ' (Modified)';
+                $uniqueIndex = 1;
+                while (Template::findOne(['name' => $uniqueName])) {
+                    $uniqueName = $template->name . ' (Modified ' . ++$uniqueIndex . ')';
+                }
+                $template->name = $uniqueName;
+                $template->save();
+
+                // Create new default template
+                $template = new Template();
             }
         }
 
-        $template = new Template();
-        $template->type = $this->type;
+        $template->type = $data['type'];
         $template->name = $data['name'];
         $template->engine = $data['engine'] ?? 'twig';
         $template->description = $data['description'] ?? '';
         $template->source = $data['source'] ?? '';
         $template->allow_for_spaces = $data['allow_for_spaces'] ?? false;
+        $template->is_default = $data['is_default'] ?? false;
 
         $this->template = $this->saveRecord($template);
 
@@ -118,74 +178,79 @@ class ImportService
 
     private function importElement(array $data): ?TemplateElement
     {
-        $element = new TemplateElement();
-        $element->setScenario(TemplateElement::SCENARIO_CREATE);
+        $element = TemplateElement::findOne([
+            'template_id' => $this->template->id,
+            'name' => $data['name'],
+        ]);
+
+        if ($element) {
+            $element->setScenario(TemplateElement::SCENARIO_EDIT_ADMIN);
+        } else {
+            $element = new TemplateElement();
+            $element->setScenario(TemplateElement::SCENARIO_CREATE);
+        }
+
         $element->template_id = $this->template->id;
         $element->name = $data['name'] ?? '';
         $element->content_type = $data['content_type'] ?? '';
         $element->title = $data['title'] ?? '';
         $element->dyn_attributes = $data['dyn_attributes'] ?? '';
+        if (is_array($element->dyn_attributes)) {
+            $element->dyn_attributes = json_encode($element->dyn_attributes);
+        }
+
+        if (!class_exists($element->content_type)) {
+            $this->addError('Element content class "' . $element->content_type . '" does not exist!');
+            return null;
+        }
 
         if (!$this->saveRecord($element)) {
             return null;
         }
 
-        if (!isset($data['elementContent'])) {
-            $this->addError('Missed content for element with name "' . $data['name'] . '"!');
-            return null;
+        if (isset($data['elementContent'])) {
+            $data['elementContent']['element_id'] = $element->id;
+            $this->importElementContent($element, $data['elementContent']);
         }
-
-        $data['elementContent']['element_id'] = $element->id;
-
-        $this->createObjectByData($element->content_type, $data['elementContent']);
 
         return $element;
     }
 
-    private function createObjectByData(string $class, array $data): ?ActiveRecord
+    private function importElementContent(TemplateElement $element, array $data): ?ActiveRecord
     {
-        if (!class_exists($class)) {
-            $this->addError('Wrong object class "' . $class . '"!');
-            return null;
-        }
-
-        try {
-            /* @var ActiveRecord $object */
-            $object = Yii::createObject($class);
-        } catch (InvalidConfigException $e) {
-            $this->addError('Cannot init object class "' . $class . '"!');
-            return null;
-        }
+        $elementContent = $element->getDefaultContent(true);
 
         foreach ($data as $name => $value) {
-            if ($name === 'id' || ($name !== 'dyn_attributes' && is_array($value))) {
+            if ($name === 'id' ||
+                ($name !== 'dyn_attributes' && is_array($value)) ||
+                !$elementContent->hasAttribute($name)) {
                 continue;
             }
-            $object->$name = $value;
+            $elementContent->$name = $value;
         }
 
-        $object = $this->saveRecord($object);
-        if (!$object) {
+        $elementContent = $this->saveRecord($elementContent);
+        if (!$elementContent) {
             return null;
         }
 
         if (isset($data['attachedFiles']) && is_array($data['attachedFiles'])) {
-            $this->attachFiles($object, $data['attachedFiles']);
+            $this->attachFiles($elementContent, $data['attachedFiles']);
         }
 
-        return $object;
+        return $elementContent;
     }
 
     private function attachFiles(ActiveRecord $record, array $files)
     {
-        $recordAttributes = $record->attributes;
         $updateRecord = false;
 
         $newFiles = [];
         foreach ($files as $fileData) {
             $file = new FileContent();
             foreach ($fileData as $attribute => $value) {
-                if (in_array($attribute, ['guid', 'object_model', 'object_id'])) {
+                if (in_array($attribute, ['guid', 'object_model', 'object_id']) ||
+                    !$file->hasAttribute($attribute)) {
                     continue;
                 }
                 if ($attribute === 'base64Content') {
@@ -201,7 +266,7 @@ class ImportService
                 $newGuid = '';
             }
 
-            foreach ($recordAttributes as $attribute => $value) {
+            foreach ($record->attributes as $attribute => $value) {
                 if ($value === $fileData['guid']) {
                     $record->$attribute = $newGuid;
                     $updateRecord = true;
@@ -216,5 +281,32 @@ class ImportService
         if ($updateRecord) {
             $record->save();
         }
+    }
+
+    public function importDefaultTemplates(): bool
+    {
+        if (!$this->getModule()) {
+            // Check because it may be called from other external module
+            return true;
+        }
+
+        $event = new DefaultTemplateEvent();
+        $event->addPath('@custom_pages/resources/templates');
+        DefaultTemplateEvent::trigger($this, self::EVENT_DEFAULT_TEMPLATES, $event);
+
+        $this->allowUpdateDefaultTemplates = true;
+
+        $result = true;
+        foreach ($event->getPaths() as $path) {
+            $result = $this->importFromFolder(Yii::getAlias($path)) && $result;
+        }
+
+        return $result;
+    }
+
+    private function getModule(): ?Module
+    {
+        $module = Yii::$app->getModule('custom_pages');
+        return $module instanceof Module && $module->isEnabled ? $module : null;
     }
 }