+++ modules/template/views/admin/editSource.php
@@ -3,8 +3,10 @@
 use humhub\modules\custom_pages\modules\template\assets\SourceEditorAsset;
 use humhub\modules\custom_pages\modules\template\models\Template;
 use humhub\modules\custom_pages\modules\template\services\ElementTypeService;
+use humhub\modules\custom_pages\modules\template\widgets\TemplateContentTable;
 use humhub\modules\custom_pages\widgets\AdminMenu;
 use humhub\modules\ui\form\widgets\ActiveForm;
+use humhub\widgets\Button;
 use yii\helpers\Html;
 use yii\helpers\Url;
 
@@ -65,7 +67,11 @@ $this->registerJsConfig('custom_pages.template.source', [
         ])->label(false); ?>
 
         <div class="clearfix">
-            <?= Html::submitButton(Yii::t('CustomPagesModule.base', 'Save'), ['class' => 'btn btn-primary', 'data-ui-loader' => ""]); ?>
+            <?= $model->canEdit() ? Button::save()->submit() : '' ?>
+            <?= $model->isNewRecord ? '' : Button::defaultType(Yii::t('CustomPagesModule.template', 'Copy'))
+                ->icon('copy')
+                ->link(Url::toRoute(['copy', 'id' => $model->id])) ?>
+            <?php if ($model->canEdit()) : ?>
             <div class="dropdown pull-right">
                 <button data-action-click="ui.modal.load" data-action-data-type="json" data-action-url="<?= Url::to(['/custom_pages/template/admin/edit-multiple', 'id' => $model->id]) ?>" class="btn btn-primary">
                     <i aria-hidden="true" class="fa fa-pencil"></i>
@@ -86,10 +92,12 @@ $this->registerJsConfig('custom_pages.template.source', [
                     <?php endforeach; ?>
                 </ul>
             </div>
+            <?php endif; ?>
         </div>
-        <br />
+        <br>
+
         <?php ActiveForm::end(); ?>
 
-        <?= \humhub\modules\custom_pages\modules\template\widgets\TemplateContentTable::widget(['template' => $model]) ?>
+        <?= TemplateContentTable::widget(['template' => $model]) ?>
     </div>
 </div>