+++ modules/template/widgets/views/templateElementAdminRow.php
@@ -1,10 +1,9 @@
 <?php
 
+use humhub\widgets\Button;
 use yii\helpers\Html;
 
 /* @var $model humhub\modules\custom_pages\modules\template\models\TemplateElement */
-
-use yii\helpers\Url;
 ?>
 <tr data-template-element-definition="<?= $model->id ?>" >
     <td class="text-nowrap">
@@ -23,22 +22,22 @@ use yii\helpers\Url;
                 <span class="label btn-success"><?= Yii::t('CustomPagesModule.base', 'Default') ?></span>
             </small>
         <?php endif; ?>
-        <?php if ($saved) : ?> 
-            <?= \humhub\widgets\DataSaved::widget() ?>
-        <?php endif; ?>
     </td>
 
     <td>
-        <a data-action-click="ui.modal.load" data-action-data-type="json" data-action-url="<?= Url::to(['/custom_pages/template/admin/edit-element', 'id' => $model->id]); ?>" class="btn btn-primary btn-icon-only btn-xs tt" href="#">
-            <i class="fa fa-pencil"></i>
-        </a> 
-        <a data-action-click="deleteElementSubmit" 
-           data-action-url="<?= Url::to(['/custom_pages/template/admin/delete-element', 'id' => $model->id]); ?>"
-           data-action-confirm="<?= Yii::t('CustomPagesModule.template', 'Do you really want to delete this element? <br />The deletion will affect all pages using this template.') ?>" 
-           data-action-confirm-header="<?= Yii::t('CustomPagesModule.template', '<strong>Confirm</strong> element deletion') ?>"
-           data-action-confirm-text="<?= Yii::t('CustomPagesModule.base', 'Delete') ?>"
-           class="btn btn-danger btn-icon-only btn-xs tt" href="#">
-            <i class="fa fa-times"></i>
-        </a>
+    <?php if ($model->template->canEdit()) : ?>
+        <?= Button::primary()->icon('pencil')->xs()
+            ->action('ui.modal.load', ['/custom_pages/template/admin/edit-element', 'id' => $model->id]) ?>
+        <?= Button::danger()->icon('times')->xs()
+            ->action('deleteElementSubmit', ['/custom_pages/template/admin/delete-element', 'id' => $model->id])
+            ->confirm(
+                Yii::t('CustomPagesModule.template', '<strong>Confirm</strong> element deletion'),
+                Yii::t('CustomPagesModule.template', 'Do you really want to delete this element? <br />The deletion will affect all pages using this template.'),
+                Yii::t('CustomPagesModule.base', 'Delete'),
+            ) ?>
+    <?php else : ?>
+        <?= Button::info()->icon('eye')->xs()
+            ->action('ui.modal.load', ['/custom_pages/template/admin/edit-element', 'id' => $model->id]) ?>
+    <?php endif; ?>
     </td>
 </tr>