+++ tests/codeception/unit/template/TemplateElementTest.php
@@ -145,6 +145,7 @@ class TemplateElementTest extends HumHubDbTestCase
 
     public function testDeleteTemplate()
     {
+        $this->becomeUser('Admin');
         $content = new RichtextElement();
         $content->content = '<p>Non Default</p>';